JS removed
Following the addition of a new JavaScript file to the repository which contains logic for the image changes, the index.html file no longer needs JavaScript embedded, and in fact needs changes in order to work with the new JavaScript file. Those changes have been made to the index.html file as part of this commit.